Move SyntaxTree stuff to ast.rs
